+++ packages/sui-widget-embedder/compiler/production.js
@@ -1,17 +1,10 @@
 const path = require('path')
 const webpack = require('webpack')
 const prodConfig = require('@s-ui/bundler/webpack.config.prod')
-
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin')
+const uglifyJsPlugin = require('@s-ui/bundler/shared/uglify')
+const {pipe, removePlugin} = require('./utils')
 const MAIN_ENTRY_POINT = './index.js'
-
-const pipe = (...funcs) => arg =>
-  funcs.reduce((value, func) => func(value), arg)
-const removePlugin = name => plugins => {
-  const pos = plugins
-    .map(p => p.constructor.toString())
-    .findIndex(string => string.match(name))
-  return [...plugins.slice(0, pos), ...plugins.slice(pos + 1)]
-}
 
 const requireOrDefault = path => {
   try {
@@ -21,9 +14,9 @@ const requireOrDefault = path => {
   }
 }
 
-module.exports = ({page, remoteCdn, globalConfig}) => {
+module.exports = ({page, remoteCdn, globalConfig = {}}) => {
   const config = requireOrDefault(
-    path.resolve(process.cwd(), 'widgets', page, 'package')
+    path.resolve(process.cwd(), 'pages', page, 'package')
   )
 
   const entry = {app: MAIN_ENTRY_POINT}
@@ -32,7 +25,7 @@ module.exports = ({page, remoteCdn, globalConfig}) => {
   }
   return webpack({
     ...prodConfig,
-    context: path.resolve(process.cwd(), 'widgets', page),
+    context: path.resolve(process.cwd(), 'pages', page),
     resolve: {
       ...prodConfig.resolve,
       alias: globalConfig.alias
@@ -43,7 +36,19 @@ module.exports = ({page, remoteCdn, globalConfig}) => {
       path: path.resolve(process.cwd(), 'public', page),
       publicPath: remoteCdn
         ? `${remoteCdn}/${page}/`
-        : prodConfig.output.publicPath
+        : prodConfig.output.publicPath,
+      jsonpFunction: `webpackJsonp-${page}`
+    },
+    optimization: {
+      ...prodConfig.optimization,
+      minimizer: [
+        uglifyJsPlugin,
+        new OptimizeCSSAssetsPlugin({
+          cssProcessorOptions: {
+            zindex: false
+          }
+        })
+      ]
     },
     plugins: pipe(
       removePlugin('HtmlWebpackPlugin'),